backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
SQLAlchemy>=2.0.36
PyMySQL>=1.1.1
python-dotenv>=1.0.1
pandas>=2.2.3
openpyxl>=3.1.5
python-multipart>=0.0.18
pydantic>=2.10.3
PyJWT>=2.9.0
requests>=2.32.3
bcrypt>=4.2.0
